+++ hwtvar.h
@@ -36,7 +36,7 @@ struct trace_context {
 	struct pmcstat_process *pp;
 	struct hwt_record_user_entry *records;
 	void *base;
-	int bufsize;
+	size_t bufsize;
 	int pid;
 	int fd;
 	int thr_fd;
